requirements.txt
fastapi[standard]>=0.135.1
sqlalchemy>=2.0.48
psycopg2-binary>=2.9.11
python-decouple>=3.8
bcrypt>=5.0.0
pyjwt>=2.11.0
numpy>=1.26,<3.0
pandas>=2.1,<3.0
openpyxl>=3.1.5
matplotlib>=3.10.8
seaborn>=0.13.2
scikit-learn>=1.6,<1.7
lightning>=2.5,<2.7
pytorch-forecasting>=1.4,<1.7
torch>=2.6,<2.11
